requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.8.0
APScheduler==3.11.0
astral==3.2
asttokens==3.0.0
asyncua==1.1.5
attrs==24.3.0
Authlib==1.4.0
av==12.3.0
backoff==2.2.1
blinker==1.9.0
boto3==1.35.60
botocore==1.35.95
branca==0.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
executing==2.1.0
fastapi==0.110.3
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.12.23
folium==0.14.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
geocoder==1.38.1
google-crc32c==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.36.1
inference==0.32.0
influxdb==5.3.2
influxdb3-python==0.9.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
marshmallow==3.24.2
matplotlib==3.8.1
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.1
nvidia-ml-py==12.560.30
onnxruntime==1.19.2
openai==1.59.5
opencv-python==4.10.0.84
packaging==24.1
pandas==2.1.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.0.0
pybase64==1.0.2
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydot==2.0.0
pyee==12.1.1
Pygments==2.18.0
PyJWT==2.10.1
pylibsrtp==0.10.0
pymongo==4.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyreadline3==3.5.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ratelim==0.1.6
reactivex==4.0.4
redis==5.0.8
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.0.1
s3transfer==0.10.4
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.37.2
structlog==24.4.0
supervision==0.25.1
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.12.12
tldextract==5.1.3
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.3.8
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.58
ultralytics-thop==2.0.13
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.0
yarl==1.18.3
zxing-cpp==2.2.0
